jxl-render: Refactor render functions
