+++ crates/jxl-render/src/vardct/mod.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use jxl_frame::{data::{LfGroup, LfGlobal, HfGlobal}, FrameHeader};
+use jxl_frame::{data::{LfGroup, LfGlobal, HfGlobal, GlobalModular, PassGroupParamsVardct, PassGroupParams}, FrameHeader};
 use jxl_grid::{CutGrid, SimpleGrid, Grid};
 use jxl_image::ImageHeader;
 use jxl_modular::ChannelShift;
@@ -12,7 +12,15 @@ use jxl_vardct::{
     BlockInfo,
 };
 
-use crate::{dct, region::ImageWithRegion, Region};
+use crate::{
+    dct,
+    region::ImageWithRegion,
+    Region,
+    IndexedFrame,
+    inner::{Reference, RenderCache},
+    Error,
+    Result, modular,
+};
 
 mod transform;
 pub use transform::transform;
@@ -30,6 +38,321 @@ use aarch64 as impls;
 mod generic;
 #[cfg(not(any(target_arch = "x86_64", target_arch = "aarch64")))]
 use generic as impls;
+
+pub fn render_vardct(
+    frame: &IndexedFrame,
+    lf_frame: Option<Reference>,
+    cache: &mut RenderCache,
+    region: Region,
+) -> Result<(ImageWithRegion, GlobalModular)> {
+    let span = tracing::span!(tracing::Level::TRACE, "Render VarDCT");
+    let _guard = span.enter();
+
+    let frame_header = frame.header();
+
+    let jpeg_upsampling = frame_header.jpeg_upsampling;
+    let shifts_cbycr: [_; 3] = std::array::from_fn(|idx| {
+        ChannelShift::from_jpeg_upsampling(jpeg_upsampling, idx)
+    });
+    let subsampled = jpeg_upsampling.into_iter().any(|x| x != 0);
+
+    let lf_global = if let Some(x) = &cache.lf_global {
+        x
+    } else {
+        let lf_global = frame.try_parse_lf_global().ok_or(Error::IncompleteFrame)??;
+        cache.lf_global = Some(lf_global);
+        cache.lf_global.as_ref().unwrap()
+    };
+    let mut gmodular = lf_global.gmodular.clone();
+    let lf_global_vardct = lf_global.vardct.as_ref().unwrap();
+
+    let width = frame_header.color_sample_width() as usize;
+    let height = frame_header.color_sample_height() as usize;
+    let (width_rounded, height_rounded) = {
+        let mut bw = (width + 7) / 8;
+        let mut bh = (height + 7) / 8;
+        let h_upsample = jpeg_upsampling.into_iter().any(|j| j == 1 || j == 2);
+        let v_upsample = jpeg_upsampling.into_iter().any(|j| j == 1 || j == 3);
+        if h_upsample {
+            bw = (bw + 1) / 2 * 2;
+        }
+        if v_upsample {
+            bh = (bh + 1) / 2 * 2;
+        }
+        (bw * 8, bh * 8)
+    };
+
+    let aligned_region = region.container_aligned(frame_header.group_dim());
+    let aligned_lf_region = {
+        // group_dim is multiple of 8
+        let aligned_region_div8 = Region {
+            left: aligned_region.left / 8,
+            top: aligned_region.top / 8,
+            width: aligned_region.width / 8,
+            height: aligned_region.height / 8,
+        };
+        if frame_header.flags.skip_adaptive_lf_smoothing() {
+            aligned_region_div8
+        } else {
+            aligned_region_div8.pad(1)
+        }.container_aligned(frame_header.group_dim())
+    };
+
+    let modular_region = modular::compute_modular_region(frame_header, &gmodular, aligned_region);
+    let modular_lf_region = modular::compute_modular_region(frame_header, &gmodular, aligned_lf_region)
+        .intersection(Region::with_size(width_rounded as u32 / 8, height_rounded as u32 / 8));
+    let aligned_region = aligned_region.intersection(Region::with_size(width_rounded as u32, height_rounded as u32));
+    let aligned_lf_region = aligned_lf_region.intersection(Region::with_size(width_rounded as u32 / 8, height_rounded as u32 / 8));
+
+    let mut fb_xyb = ImageWithRegion::from_region(3, aligned_region);
+    let fb_stride = aligned_region.width as usize;
+
+    let lf_groups = &mut cache.lf_groups;
+    tracing::trace_span!("Load LF groups").in_scope(|| {
+        crate::load_lf_groups(frame, lf_global, lf_groups, modular_lf_region, &mut gmodular)
+    })?;
+
+    let group_dim = frame_header.group_dim();
+    let (hf_cfl_data, mut lf_xyb) = tracing::trace_span!("Copy LFQuant").in_scope(|| {
+        let mut hf_cfl_data = (!subsampled).then(|| {
+            ImageWithRegion::from_region(2, aligned_lf_region.downsample(3))
+        });
+
+        let mut lf_xyb = ImageWithRegion::from_region(3, aligned_lf_region);
+
+        if let Some(x) = lf_frame {
+            x.image.clone_region_channel(aligned_lf_region, 0, &mut lf_xyb.buffer_mut()[0]);
+            x.image.clone_region_channel(aligned_lf_region, 1, &mut lf_xyb.buffer_mut()[1]);
+            x.image.clone_region_channel(aligned_lf_region, 2, &mut lf_xyb.buffer_mut()[2]);
+        }
+
+        let lf_groups_per_row = frame_header.lf_groups_per_row();
+        for idx in 0..frame_header.num_lf_groups() {
+            let Some(lf_group) = lf_groups.get(&idx) else { continue; };
+
+            let lf_group_x = idx % lf_groups_per_row;
+            let lf_group_y = idx / lf_groups_per_row;
+            let left = lf_group_x * frame_header.group_dim();
+            let top = lf_group_y * frame_header.group_dim();
+            let lf_group_region = Region {
+                left: left as i32,
+                top: top as i32,
+                width: group_dim,
+                height: group_dim,
+            };
+            if aligned_lf_region.intersection(lf_group_region).is_empty() {
+                continue;
+            }
+
+            let left = left - aligned_lf_region.left as u32;
+            let top = top - aligned_lf_region.top as u32;
+
+            if lf_frame.is_none() {
+                let quantizer = &lf_global_vardct.quantizer;
+                let lf_coeff = lf_group.lf_coeff.as_ref().unwrap();
+                let channel_data = lf_coeff.lf_quant.image().channel_data();
+                let [lf_x, lf_y, lf_b] = lf_xyb.buffer_mut() else { panic!() };
+                copy_lf_dequant(
+                    lf_x,
+                    left as usize >> shifts_cbycr[0].hshift(),
+                    top as usize >> shifts_cbycr[0].vshift(),
+                    quantizer,
+                    lf_global.lf_dequant.m_x_lf,
+                    &channel_data[1],
+                    lf_coeff.extra_precision,
+                );
+                copy_lf_dequant(
+                    lf_y,
+                    left as usize >> shifts_cbycr[1].hshift(),
+                    top as usize >> shifts_cbycr[1].vshift(),
+                    quantizer,
+                    lf_global.lf_dequant.m_y_lf,
+                    &channel_data[0],
+                    lf_coeff.extra_precision,
+                );
+                copy_lf_dequant(
+                    lf_b,
+                    left as usize >> shifts_cbycr[2].hshift(),
+                    top as usize >> shifts_cbycr[2].vshift(),
+                    quantizer,
+                    lf_global.lf_dequant.m_b_lf,
+                    &channel_data[2],
+                    lf_coeff.extra_precision,
+                );
+            }
+
+            let Some(hf_meta) = &lf_group.hf_meta else { continue; };
+
+            if let Some(cfl) = &mut hf_cfl_data {
+                let corr = &lf_global_vardct.lf_chan_corr;
+                let [x_from_y, b_from_y] = cfl.buffer_mut() else { panic!() };
+                let group_x_from_y = &hf_meta.x_from_y;
+                let group_b_from_y = &hf_meta.b_from_y;
+                let left = left as usize / 8;
+                let top = top as usize / 8;
+                for y in 0..group_x_from_y.height() {
+                    for x in 0..group_x_from_y.width() {
+                        let v = *group_x_from_y.get(x, y).unwrap();
+                        let kx = corr.base_correlation_x + (v as f32 / corr.colour_factor as f32);
+                        *x_from_y.get_mut(left + x, top + y).unwrap() = kx;
+                    }
+                }
+                for y in 0..group_b_from_y.height() {
+                    for x in 0..group_b_from_y.width() {
+                        let v = *group_b_from_y.get(x, y).unwrap();
+                        let kb = corr.base_correlation_b + (v as f32 / corr.colour_factor as f32);
+                        *b_from_y.get_mut(left + x, top + y).unwrap() = kb;
+                    }
+                }
+            }
+        }
+
+        (hf_cfl_data, lf_xyb)
+    });
+
+    if lf_frame.is_none() {
+        if !subsampled {
+            tracing::trace_span!("LF CfL").in_scope(|| {
+                chroma_from_luma_lf(
+                    lf_xyb.buffer_mut(),
+                    &lf_global_vardct.lf_chan_corr,
+                );
+            });
+        }
+
+        if !frame_header.flags.skip_adaptive_lf_smoothing() {
+            tracing::trace_span!("Adaptive LF smoothing").in_scope(|| {
+                adaptive_lf_smoothing(
+                    lf_xyb.buffer_mut(),
+                    &lf_global.lf_dequant,
+                    &lf_global_vardct.quantizer,
+                );
+            });
+        }
+    }
+
+    let hf_global = if let Some(x) = &cache.hf_global {
+        Some(x)
+    } else {
+        cache.hf_global = frame.try_parse_hf_global(Some(lf_global)).transpose()?;
+        cache.hf_global.as_ref()
+    };
+
+    tracing::trace_span!("Decode pass groups").in_scope(|| -> Result<_> {
+        let Some(hf_global) = hf_global else { return Ok(()); };
+        let groups_per_row = frame_header.groups_per_row();
+        for pass_idx in 0..frame_header.passes.num_passes {
+            for group_idx in 0..frame_header.num_groups() {
+                let lf_group_idx = frame_header.lf_group_idx_from_group_idx(group_idx);
+                let Some(lf_group) = lf_groups.get(&lf_group_idx) else { continue; };
+                if lf_group.hf_meta.is_none() {
+                    continue;
+                }
+                let Some(bitstream) = frame.pass_group_bitstream(pass_idx, group_idx).transpose()? else { continue; };
+                let allow_partial = bitstream.partial;
+                let mut bitstream = bitstream.bitstream;
+
+                let group_x = group_idx % groups_per_row;
+                let group_y = group_idx / groups_per_row;
+                let left = group_x * group_dim;
+                let top = group_y * group_dim;
+                let group_width = group_dim.min(width_rounded as u32 - left);
+                let group_height = group_dim.min(height_rounded as u32 - top);
+
+                let group_region = Region {
+                    left: left as i32,
+                    top: top as i32,
+                    width: group_width,
+                    height: group_height,
+                };
+                if group_region.intersection(modular_region).is_empty() {
+                    continue;
+                }
+
+                let mut grid_xyb;
+                let vardct = if group_region.intersection(aligned_region).is_empty() {
+                    None
+                } else {
+                    let left = left - aligned_region.left as u32;
+                    let top = top - aligned_region.top as u32;
+
+                    let [fb_x, fb_y, fb_b] = fb_xyb.buffer_mut() else { panic!() };
+                    grid_xyb = [(0usize, fb_x), (1, fb_y), (2, fb_b)].map(|(idx, fb)| {
+                        let hshift = shifts_cbycr[idx].hshift();
+                        let vshift = shifts_cbycr[idx].vshift();
+                        let group_width = group_width >> hshift;
+                        let group_height = group_height >> vshift;
+                        let left = left >> hshift;
+                        let top = top >> vshift;
+                        let offset = top as usize * fb_stride + left as usize;
+                        CutGrid::from_buf(&mut fb.buf_mut()[offset..], group_width as usize, group_height as usize, fb_stride)
+                    });
+
+                    Some(PassGroupParamsVardct {
+                        lf_vardct: lf_global_vardct,
+                        hf_global,
+                        hf_coeff_output: &mut grid_xyb,
+                    })
+                };
+
+                let shift = frame.pass_shifts(pass_idx);
+                let result = jxl_frame::data::decode_pass_group(
+                    &mut bitstream,
+                    PassGroupParams {
+                        frame_header,
+                        lf_group,
+                        pass_idx,
+                        group_idx,
+                        shift,
+                        gmodular: &mut gmodular,
+                        vardct,
+                        allow_partial,
+                    },
+                );
+                if !allow_partial {
+                    result?;
+                }
+            }
+        }
+        Ok(())
+    })?;
+
+    tracing::trace_span!("Extra channel inverse transform").in_scope(|| {
+        gmodular.modular.inverse_transform();
+    });
+
+    tracing::trace_span!("Dequant HF").in_scope(|| {
+        let Some(hf_global) = hf_global else { return; };
+        dequant_hf_varblock(
+            &mut fb_xyb,
+            frame.image_header(),
+            frame_header,
+            lf_global,
+            &*lf_groups,
+            hf_global,
+        );
+    });
+
+    if let Some(cfl) = hf_cfl_data {
+        tracing::trace_span!("HF CfL").in_scope(|| {
+            if hf_global.is_none() {
+                return;
+            }
+            chroma_from_luma_hf(&mut fb_xyb, &cfl);
+        });
+    }
+
+    tracing::trace_span!("Transform varblocks").in_scope(|| {
+        transform_with_lf(
+            &lf_xyb,
+            &mut fb_xyb,
+            frame_header,
+            &*lf_groups,
+        );
+    });
+
+    Ok((fb_xyb, gmodular))
+}
 
 pub fn copy_lf_dequant(
     out: &mut SimpleGrid<f32>,
